use secret instead of stirng
