+++ core/src/main/scala/gospeak/core/services/youtube/YoutubeConf.scala
@@ -1,3 +1,5 @@
 package gospeak.core.services.youtube
 
-final case class YoutubeConf(secret: String)
+import gospeak.libs.scala.domain.Secret
+
+final case class YoutubeConf(secret: Secret)
